requirements.txt
PySide6>=6.5.0
Pillow>=10.0.0
numpy>=1.24.0
sprite-splitter>=0.0.4